Moves the auto panelization dialogs into their own files to reduce the length and complexity of JobPanel. See #501.
